Pipe Spawner
Made the pipe as prefab by dragging to project. Removed from hierarchy, then created a new empty called pipespawner and added script to it.

In the pipespawn script, I made a reference to pipe gameobject and spawned it using instantiate. The proble m now is that pipes are created every frame continuously.